Rename charm and service.
