+++ README.md
@@ -1,4 +1,7 @@
 # Overview
 
-This charm deploys the Ubuntu ESM (Extended Security Maintenance) mirroring service.
-It's meant to be deployed alongside apache2 so that the mirrored archive can be exposed.
+This subordinate charm deploys an application to mirror and periodically sync
+an Ubuntu archive.
+
+It's meant to be deployed alongside apache2 so that the mirrored archive can be
+exposed.
