--- resources/index.html
@@ -1,9 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Ubuntu ESM (Extended Security Maintenance) archive</title>
-  </head>
-  <body>
-    <h1>Ubuntu ESM (Extended Security Maintenance) archive</h1>
-  </body>
-</html>
